requirements.txt
python-dotenv==1.2.2
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
requests==2.32.5
urllib3==2.6.3
darkdetect==0.8.0
packaging==26.0
customtkinter==5.2.2
